+++ src/clj_foundationdb/core.clj
@@ -61,6 +61,16 @@
             value (.pack (Tuple/from (to-array [value])))]
         (.set tr key value)))))
 
+(defn set-tuple-vals
+  "Set a value for the collection of keys with tuple"
+  [keys value]
+  (reify
+    java.util.function.Function
+    (apply [this tr]
+      (let [keys  (map #(.pack (Tuple/from (to-array %1))) keys)
+            value (.pack (Tuple/from (to-array [value])))]
+        (doall (map #(.set tr %1 value) keys))))))
+
 (spec/fdef set-val
            :args (spec/cat :key (spec/coll-of string?) :value string?))
 
@@ -117,6 +127,27 @@
 (spec/fdef get-range-startswith
            :args (spec/cat :prefix string?))
 
+(defn watch
+  [key]
+  (reify
+    java.util.function.Function
+    (apply [this tr]
+      (let [key (.getBytes key)]
+        (.watch tr key)))))
+
+(defn get-tuple-range-startswith
+  "Get a range of key values as a vector that starts with prefix"
+  [prefix]
+  (reify
+    java.util.function.Function
+    (apply [this tr]
+      (let [prefix (.pack (Tuple/from (to-array prefix)))
+            r      (Range/startsWith prefix)]
+        (->> (.getRange tr r)
+             (mapv #(vector
+                     (bytes-to-str (.getKey %1))
+                     (bytes-to-str (.getValue %1)))))))))
+
 (defn get-range
   "Get a range of key values as a vector"
   [begin end]
@@ -133,6 +164,9 @@
 
 (spec/fdef get-range
            :args (spec/cat :begin string? :end string?))
+
+;; https://stackoverflow.com/a/21421524/2610955
+;; TODO: Unfortunately this doesn't seem to work and only retrives until keys prefixing x
 
 (defn get-all
   "Get all key values as a vector"
@@ -153,6 +187,7 @@
            :args (spec/cat :begin string? :end string?))
 
 ;; https://stackoverflow.com/a/21421524/2610955
+;; TODO: Unfortunately this doesn't seem to work and only deletes till keys prefixing x
 
 (defn clear-all
   "Clear all the keys and values"
@@ -160,3 +195,83 @@
   (let [begin ""
         end   "xFF"]
     (clear-range begin end)))
+
+(defn last-less-than
+  "Returns key and value pairs with keys less than the given key for the given limit"
+  ([key]
+   (last-less-than key 1))
+  ([key limit]
+   (reify
+     java.util.function.Function
+     (apply [this tr]
+       (let [key (KeySelector/lastLessThan (.getBytes key))
+             end (.add key limit)
+             range-query (.getRange tr key end)]
+         (->> range-query
+              (mapv #(vector
+                      (bytes-to-str (.getKey %1))
+                      (bytes-to-str (.getValue %1))))))))))
+
+(spec/fdef last-less-than
+           :args (spec/cat :key string? :limit (spec/? pos-int?))
+           :ret (spec/tuple string? number?))
+
+(defn last-less-or-equal
+  "Returns key and value pairs with keys less than or equal the given key for the given limit"
+  ([key]
+   (last-less-or-equal key 1))
+  ([key limit]
+   (reify
+     java.util.function.Function
+     (apply [this tr]
+       (let [key (KeySelector/lastLessOrEqual (.getBytes key))
+             end (.add key limit)
+             range-query (.getRange tr key end)]
+         (->> range-query
+              (mapv #(vector
+                      (bytes-to-str (.getKey %1))
+                      (bytes-to-str (.getValue %1))))))))))
+
+(spec/fdef last-less-or-equal
+           :args (spec/cat :key string? :limit (spec/? pos-int?))
+           :ret (spec/tuple string? number?))
+
+(defn first-greater-than
+  "Returns key and value pairs with keys greater than the given key for the given limit"
+  ([key]
+   (first-greater-than key 1))
+  ([key limit]
+   (reify
+     java.util.function.Function
+     (apply [this tr]
+       (let [key (KeySelector/firstGreaterThan (.getBytes key))
+             end (.add key limit)
+             range-query (.getRange tr key end)]
+         (->> range-query
+              (mapv #(vector
+                      (bytes-to-str (.getKey %1))
+                      (bytes-to-str (.getValue %1))))))))))
+
+(spec/fdef first-greater-than
+           :args (spec/cat :key string? :limit (spec/? pos-int?))
+           :ret (spec/tuple string? number?))
+
+(defn first-greater-or-equal
+  "Returns key and value pairs with keys greater than or equal to the given key for the given limit"
+  ([key]
+   (first-greater-or-equal key 1))
+  ([key limit]
+   (reify
+     java.util.function.Function
+     (apply [this tr]
+       (let [key (KeySelector/firstGreaterOrEqual (.getBytes key))
+             end (.add key limit)
+             range-query (.getRange tr key end)]
+         (->> range-query
+              (mapv #(vector
+                      (bytes-to-str (.getKey %1))
+                      (bytes-to-str (.getValue %1))))))))))
+
+(spec/fdef first-greater-or-equal
+           :args (spec/cat :key string? :limit (spec/? pos-int?))
+           :ret (spec/tuple string? number?))