+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## [Unreleased]
 
+- Add watch, set-tuple-vals, get-tuple-range-startswith, first-greater-than, first-greater-or-equal, last-less-than, last-less-or-equal
 - Add clear-key, get-range, get-range-startswith, get-all, clear-range
 - Add method to set a value, get a value, get a range of keys, clear database